% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonFunction.R, R/naive.r
\name{local_bic_gaussian}
\alias{local_bic_gaussian}
\title{Local BIC score for a Gaussian Bayesian network node}
\usage{
local_bic_gaussian(j, parents, X)

local_bic_gaussian(j, parents, X)
}
\arguments{
\item{j}{Integer, index of the child variable (column index in `X`).}

\item{parents}{Integer vector of parent indices (can be `integer(0)`).}

\item{X}{Numeric matrix or data frame of size N x p containing the data.}
}
\value{
Numeric, local BIC score contribution for node \code{j}.

Numeric, local BIC score contribution for node \code{j}.
}
\description{
Local BIC score for a Gaussian Bayesian network node

Local BIC score for a Gaussian Bayesian network node
}
