% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonFunction.R, R/naive.r
\name{local_bic_multinomial}
\alias{local_bic_multinomial}
\title{Local BIC score for a multinomial (discrete) Bayesian network node}
\usage{
local_bic_multinomial(j, parents, X)

local_bic_multinomial(j, parents, X)
}
\arguments{
\item{j}{Integer, index of the child variable (column index in `X`).}

\item{parents}{Integer vector of parent indices (can be `integer(0)`).}

\item{X}{Data frame or matrix of size N x p with categorical variables
(will be coerced to factors internally).}
}
\value{
Numeric, local BIC score contribution for node \code{j}.

Numeric, local BIC score contribution for node \code{j}.
}
\description{
Local BIC score for a multinomial (discrete) Bayesian network node

Local BIC score for a multinomial (discrete) Bayesian network node
}
